chore: move control panel box component into shared folder
