+++ src/components/shared/ControlPanelBox/ControlPanelBox.styles.js
